replay viewer: add 1/2x playback speed; Space pause verified on the shipped page (#2)
The replay transport gains a 0.5x speed: command '5' selects the
ReplayHalfSpeed sentinel (speed == 0) on the viewer state,
advanceReplayFrame spends one tick every other frame via a halfPhase
parity bit toggled as its first statement, '-' now floors at 1/2x
(1 div 2 == 0) and '+' climbs back to 1x. The chrome's "sp" field
carries the float displaySpeed() -- 0.5 while at half -- and
WireConstantsJs prepends 0.5 to the emitted speeds array, so
chrome_common builds a 0.5x chip mapped to command '5' (its raw
file:// fallbacks carry 0.5 too). The live server path reports 1.0.

This repo ships exactly one viewer page (client/replay_broadcast.html
-> dist/index.html) and no league shell: that page already binds Space
(preventDefault + send ' ') and relays the '1'..'9' digit range down
the command channel, so the new '5' works with zero page edits, and
the platform shell that embeds it already forwards transport commands
by postMessage.

tests: test_replay grows a half-speed block (sentinel, 0.5 display and
chrome sp, tick-every-other-frame parity, the '-'/'+' floor and exit);
test_broadcast pins speeds:[0.5,1,2,3,4,8,16] in WireConstantsJs.

+++ src/ecos/broadcast.nim
@@ -22,7 +22,9 @@ type
     tick*, maxTick*, startTick*: int
     generation*, generations*, ticksPerGeneration*: int
     playing*: bool
-    speed*: int
+    speed*: float
+      ## The speed the chrome shows: 0.5 while the replay-only 1/2x
+      ## (ReplayHalfSpeed) is selected, else the integer multiplier.
     looping*, skipLulls*, fastForward*, transportEnabled*: bool
     over*: bool
     reason*, ending*: string
